Created vehicle tab widget
